# Trim TSDB head chunks after being cut, to reduce inuse memory