--- a/tsdb/chunkenc/xor.go
+++ b/tsdb/chunkenc/xor.go
@@ -49,6 +49,10 @@ import (
 	"math/bits"
 )
 
+const (
+	chunkCompactCapacityThreshold = 32
+)
+
 // XORChunk holds XOR encoded sample data.
 type XORChunk struct {
 	b bstream
@@ -73,6 +77,14 @@ func (c *XORChunk) Bytes() []byte {
 // NumSamples returns the number of samples in the chunk.
 func (c *XORChunk) NumSamples() int {
 	return int(binary.BigEndian.Uint16(c.Bytes()))
+}
+
+func (c *XORChunk) Compact() {
+	if l := len(c.b.stream); cap(c.b.stream) > l+chunkCompactCapacityThreshold {
+		buf := make([]byte, l)
+		copy(buf, c.b.stream)
+		c.b.stream = buf
+	}
 }
 
 // Appender implements the Chunk interface.