I'm currently working on a new experimental storage in Cortex based on TSDB (and Thanos). While benchmarking it and comparing the **inuse memory**, I've noticed that for the same exact series TSDB uses more memory than the Cortex chunks storage.

While profiling it I've found that part of such difference is given by an optimization we have in Cortex but doesn't exist in TSDB: chunks are not trimmed after being cut in the TSDB head.

In particular, a chunk byte slice is allocated in TSDB with an initial capacity of 128 bytes and then it double each time it's full:
https://github.com/prometheus/prometheus/blob/master/tsdb/chunkenc/xor.go#L58-L61

When a chunk is cut, the extra capacity of that `[]byte` is never reclaimed and we may end up (theoretically) wasting 2x space. It's not 2x space in the real world, but it may be significant. In the Cortex chunks storage, we check it and compact the `[]byte` if the capacity exceed 32 bytes.

I've applied the same optimization in TSDB and, according to our preliminary tests, it has reduced the memory used by TSDB by about -11% in our test cluster. More accurate benchmarks are required.

The change proposed in this PR is a trade-off between inuse memory and allocations: we're doing more allocations to reduce the inuse memory. For example, running the `BenchmarkLoadWAL` with `-benchmem` we can see:

```
benchmark                                                                  old ns/op      new ns/op      delta
BenchmarkLoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200-4     686987166      697524794      +1.53%
BenchmarkLoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50-4     1113564403     1128349782     +1.33%
BenchmarkLoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480-4     481137676      476783942      -0.90%

benchmark                                                                  old allocs     new allocs     delta
BenchmarkLoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200-4     640057         710074         +10.94%
BenchmarkLoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50-4     2395984        2395953        -0.00%
BenchmarkLoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480-4     569179         609208         +7.03%

benchmark                                                                  old bytes     new bytes     delta
BenchmarkLoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200-4     57155592      71857912      +25.72%
BenchmarkLoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50-4     365385744     364694736     -0.19%
BenchmarkLoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480-4     67557674      75518242      +11.78%
```