@codesome 2020-01-24
Some visible differences in resident memory.

Before merging I would also like to verify if it has any potential race. Consider the following cases
1. `Compact()` is called when a query is iterating that chunk - this might not be a problem as we take a subslice for the iterator already.
2. A similar case to above, but `Compact()` is called when the query is trying to get the iterator for the chunk. Here would the swapping of the byte stream in `Compact()` and taking the subslice in the iterator method cause any race?