@pracucci 2020-01-27
Thanks @codesome for the feedback. I'm going to close this PR, cause an extended test running for a couple of days showed no significant benefit on the long run for our use case. We're still seeing TSDB significantly using more memory than Cortex chunks storage for the same series, but we need to spend more time on that in order to correctly pinpoint the causes.

_Thanks for your time so far._