@pracucci 2020-01-31
I'm re-opening this PR cause further tests actually showed some benefits in the inuse memory reduction. Looking again at the prombench metrics, the same improvements can be observed there, except that it's less than initially estimated (-11%).

The actual inuse memory reduction is in the range of -3%,-6% when we look at the entire process (in this case Prometheus) and not just the chunks.

## Metrics from Prombench

RSS memory:

![rss](https://user-images.githubusercontent.com/1701904/73549491-f78d8a00-4442-11ea-817b-f7e0676b0653.png)

Allocated memory picking the max over 5m:
`max_over_time(go_memstats_alloc_bytes{job="prometheus",namespace="prombench-[[pr-number]]",prometheus=~"(test+).*"}[5m])`

![allocated](https://user-images.githubusercontent.com/1701904/73549504-fc523e00-4442-11ea-8993-82462842b124.png)

Heap inuse picking the max over 5m:
`max_over_time(go_memstats_heap_inuse_bytes{job="prometheus",namespace="prombench-[[pr-number]]",prometheus=~"(test+).*"}[5m])`

![heap](https://user-images.githubusercontent.com/1701904/73549537-11c76800-4443-11ea-8b2c-18ec842c3aa6.png)

## Other considerations

An argument against this PR is that we may `Compact()` all chunks nearly at the same time, in case of the same scrape interval for all metrics and not churn. In general, this PR obviously increase memory allocations.

The following shows allocations with `rate()` over `[5m]` window to see more smooth metrics which are easier to compare:

![allocations](https://user-images.githubusercontent.com/1701904/73550325-b1392a80-4444-11ea-963d-88e66d78267b.png)

## What's next?

I would like to *get more feedback* and *more critical eyes* on the possible positive/negative effects of such change (ie. are we OK having more memory allocations to save a bit on inuse memory?).