@pracucci 2020-01-31
> Before merging I would also like to verify if it has any potential race. Consider the following cases
> 
> 1. `Compact()` is called when a query is iterating that chunk - this might not be a problem as we take a subslice for the iterator already.
> 2. A similar case to above, but `Compact()` is called when the query is trying to get the iterator for the chunk. Here would the swapping of the byte stream in `Compact()` and taking the subslice in the iterator method cause any race?

_I'm not much familiar with all the TSDB internals, so I may miss something._

To my understanding, a query gets an Iterator from a `safeChunk`, which wraps it around the `memSeries.Lock()`:
https://github.com/prometheus/prometheus/blob/1ebcf3d5e479a5d78e8b79ee4425c3b586e17197/tsdb/head.go#L1282-L1287

Likewise, when samples are appended in the `Commit()`, the append is wrapped around the `memSeries.Lock()`:
https://github.com/prometheus/prometheus/blob/1ebcf3d5e479a5d78e8b79ee4425c3b586e17197/tsdb/head.go#L1008-L1011

The `chunk.Compact()` is called in the `memSeries.cut()`, which is called only within the `memSeries.append()` which is wrapped by the lock, so it looks safe to me.