@pracucci 2020-02-01
## Metrics from Cortex ingesters with blocks storage

We now have some data points to compare inuse memory in Cortex ingesters (with the experimental blocks storage based on TSDB) with and without this optimization. It looks we get similar results as the Prombench.

<img width="1419" alt="heap_inuse" src="https://user-images.githubusercontent.com/1701904/73588088-074db280-44c5-11ea-8d58-8c02f7b78906.png">

<img width="1425" alt="alloc_bytes" src="https://user-images.githubusercontent.com/1701904/73588090-087edf80-44c5-11ea-8d59-cb7e9d837945.png">