@bwplotka 2020-02-05
I understand this will be irrelevant once @codesome PR will be merged, but I really like this in the interface as a nice practice that we should do in other places. 

(But maybe making code bit more complex just for this purpose is not worth it)